Fixed OSX C++11 compilation warnings
git-svn-id: svn://svn.code.sf.net/p/glmixer/Source/trunk@1420 a4deb95c-d308-4ec4-8dc1-6419bbeeb61a

+++ src/RenderingManager.cpp
@@ -1675,7 +1675,7 @@ double RenderingManager::getAvailableDepthFrom(double depth) const {
     if (foundDepth > MAX_DEPTH_LAYER) {
 
         // compute the depth between sources to uniformly distribute them
-        double d_inc = MAX_DEPTH_LAYER / (float)(_front_sources.size() + dropBasket.size() + 1 );
+        double d_inc = MAX_DEPTH_LAYER / (double)(_front_sources.size() + dropBasket.size() + 1 );
 
         // if this is not a problem for distinguishing depth
         if (d_inc > DEPTH_EPSILON)
